some progress, and there was much refactoring
Now failing on running a pipeline

+++ spec/support/git_repository.rb
@@ -14,7 +14,7 @@ class GitRepository
   end
 
   def destroy
-    FileUtils.remove_entry_secure(@path)
+    FileUtils.remove_entry_secure(@path) if @path
   end
 
   def create_bad_commit
